+++ api/model/database/categories.py
@@ -1,0 +1,9 @@
+from api import db
+
+class Categoria(db.Model):
+    __tablename__ = 'categorias'
+    id = db.Column(db.Integer, primary_key=True)
+    nome = db.Column(db.String(80), unique=True, nullable=False)
+
+    def __init__(self, nome):
+        self.nome = nome
