+++ api/model/database/posts.py
@@ -15,11 +15,3 @@ class Postagem(db.Model):
         self.texto = texto
         self.criador = criador
         self.categoria = categoria
-
-class Categoria(db.Model):
-    __tablename__ = 'categorias'
-    id = db.Column(db.Integer, primary_key=True)
-    nome = db.Column(db.String(80), unique=True, nullable=False)
-
-    def __init__(self, nome):
-        self.nome = nome
